refactor: move metadata-table to correct folder
