+++ src/app/compare/compare.component.ts
@@ -4,7 +4,7 @@ import { CompareData } from './compare-data';
 import { DiffEditorModel, MonacoEditorModule } from 'ngx-monaco-editor-v2';
 import { TabService } from '../shared/services/tab.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { MetadataTableComponent } from '../shared/components/display-table/metadata-table.component';
+import { MetadataTableComponent } from '../shared/components/metadata-table/metadata-table.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TitleCasePipe } from '@angular/common';
 import { NodeLinkStrategy, nodeLinkStrategyConst } from '../shared/enums/node-link-strategy';
