+++ src/app/shared/components/metadata-table/metadata-table.component.html
